PhaseList change; Changed to member variables from int to size_t. (numVolPhases_ and numSurPhases_).

+++ Electrode/src/Electrode_Capacity.cpp
@@ -425,8 +425,8 @@ void Electrode::setCapacityCoeff_LiSi() const
 
     double  capacityLeftSpeciesCoeffLi13Si4 =  DoDElectronTmp;
 
-    for (int iph = 0; iph < NumVolPhases_; iph++) {
-        if (iph == solnPhase_ || iph == metalPhase_) {
+    for (size_t iph = 0; iph < NumVolPhases_; iph++) {
+        if (iph == (size_t) solnPhase_ || iph == (size_t) metalPhase_) {
             continue;
         }
 
@@ -540,8 +540,8 @@ void Electrode::setCapacityCoeff_LiSi_Li() const
 
     double  capacityLeftSpeciesCoeffLi13Si4 =  DoDElectronTmp;
 
-    for (int iph = 0; iph < NumVolPhases_; iph++) {
-        if (iph == solnPhase_ || iph == metalPhase_) {
+    for (size_t iph = 0; iph < NumVolPhases_; iph++) {
+        if (iph == (size_t) solnPhase_ || iph == (size_t) metalPhase_) {
             continue;
         }
 
@@ -607,8 +607,8 @@ void Electrode::setCapacityCoeff_FeS2() const
 
     double  capacityLeftSpeciesCoeffFeS2 =  DoDElectronTmp;
 
-    for (int iph = 0; iph < NumVolPhases_; iph++) {
-        if (iph == solnPhase_ || iph == metalPhase_) {
+    for (size_t iph = 0; iph < NumVolPhases_; iph++) {
+        if (iph == (size_t) solnPhase_ || iph == (size_t) metalPhase_) {
             continue;
         }
 
@@ -770,8 +770,8 @@ void Electrode::setCapacityCoeff_FeS2_Combo() const
 
     double  capacityLeftSpeciesCoeffFeS2 =  DoDElectronTmp;
 
-    for (int iph = 0; iph < NumVolPhases_; iph++) {
-        if (iph == solnPhase_ || iph == metalPhase_) {
+    for (size_t iph = 0; iph < NumVolPhases_; iph++) {
+        if (iph == (size_t) solnPhase_ || iph == (size_t) metalPhase_) {
             continue;
         }
 
@@ -857,8 +857,8 @@ void Electrode::setCapacityCoeff_MCMB() const
 {
     double  capacityLeftSpeciesCoeff=  1.0;
 
-    for (int iph = 0; iph < NumVolPhases_; iph++) {
-        if (iph == solnPhase_ || iph == metalPhase_) {
+    for (size_t iph = 0; iph < NumVolPhases_; iph++) {
+        if (iph == (size_t) solnPhase_ || iph == (size_t) metalPhase_) {
             continue;
         }
